+++ internal/app/shell.go
@@ -256,6 +256,7 @@ func Shell(props ShellProps) uic.Node {
 		uic.CreateElement(TxnEditHost),
 		uic.CreateElement(TxnColumnsHost),
 		uic.CreateElement(TxnSmartCatHost),
+		uic.CreateElement(TxnLinkHost),
 		uic.CreateElement(AccountEditHost),
 		uic.CreateElement(BudgetEditHost),
 		uic.CreateElement(GoalEditHost),